requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
requests>=2.31.0
httpx>=0.25.0
numpy>=1.24.0
--extra-index-url https://download.pytorch.org/whl/nightly/cpu
torch>=2.2.0
torchaudio>=2.2.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0 